// suiviDeploiementMobiliers/src/app/core/models/summary.ts
export interface Summary {

}
export interface RegionSummary {
  name: string;
  delivered: number;
  enCours: number;
  total: number;
}

export interface PersonnelSummary {
  name: string;
  total: number;
  delivered: number;
  percentage: number;
}

export interface MobilierSummary {
  name: string;
  delivered: number;
  enCours: number;
  total: number;
}